+++ cli/src/main.rs
@@ -15,39 +15,44 @@ use std::{
 
 use anyhow::{anyhow, Context};
 use clap::{CommandFactory, Parser};
-use clap_complete::Generator;
+use clap_complete::aot::generate;
+use flexi_logger::AdaptiveFormat;
 use ignore::{overrides::OverrideBuilder, types::TypesBuilder, WalkBuilder};
-use log::error;
-use rayon::iter::ParallelBridge;
+use log::{error, info};
+use parse::parallel_parse;
 use typeshare_core::language::GenericConstraints;
 #[cfg(feature = "go")]
 use typeshare_core::language::Go;
 #[cfg(feature = "python")]
 use typeshare_core::language::Python;
 use typeshare_core::{
+    context::ParseContext,
     language::{CrateName, Kotlin, Language, Scala, SupportedLanguage, Swift, TypeScript},
     parser::ParsedData,
+    reconcile::reconcile_aliases,
 };
 
 use crate::{
     args::{Args, Command},
     config::Config,
-    parse::{all_types, parse_input, parser_inputs},
+    parse::all_types,
     writer::{write_generated, Output},
 };
 
 fn main() -> anyhow::Result<()> {
-    flexi_logger::Logger::try_with_env()
-        .unwrap()
-        .start()
-        .unwrap();
+    flexi_logger::Logger::try_with_env_or_str("info")?
+        .adaptive_format_for_stderr(AdaptiveFormat::Opt)
+        .adaptive_format_for_stdout(AdaptiveFormat::Opt)
+        .start()?;
 
     let options = Args::parse();
 
     if let Some(options) = options.subcommand {
         match options {
             Command::Completions { shell } => {
-                shell.generate(&Args::command(), &mut io::stdout().lock())
+                let mut cmd = Args::command();
+                let bin_name = cmd.get_name().to_string();
+                generate(shell, &mut cmd, bin_name, &mut io::stdout());
             }
         }
 
@@ -59,15 +64,25 @@ fn main() -> anyhow::Result<()> {
     let config_file = options.config_file.as_deref();
 
     if options.output.generate_config {
-        let config = override_configuration(Config::default(), &options)?;
-        config::store_config(&config, config_file).context("Failed to create new config file")?;
-        return Ok(());
+        override_configuration(Config::default(), &options)
+            .and_then(|config| config::store_config(&config, config_file))
+            .inspect_err(|err| error!("typeshare failed to create new config file: {err}"))
+    } else {
+        generate_types(config_file, &options).inspect_err(|err| {
+            error!("typeshare failed to generate types: {err}");
+        })
     }
+}
+
+fn generate_types(config_file: Option<&Path>, options: &Args) -> anyhow::Result<()> {
+    info!("typeshare started generating types");
 
     let config = config::load_config(config_file).context("Unable to read configuration file")?;
-    let config = override_configuration(config, &options)?;
+    let config = override_configuration(config, options)?;
 
     let directories = options.directories.as_slice();
+
+    info!("Using directories: {directories:?}");
 
     let language_type = match options.language {
         None => panic!("no language specified; `clap` should have guaranteed its presence"),
@@ -83,36 +98,6 @@ fn main() -> anyhow::Result<()> {
         },
     };
 
-    let mut types = TypesBuilder::new();
-    types
-        .add("rust", "*.rs")
-        .context("Failed to add rust type extensions")?;
-    types.select("rust");
-
-    // This is guaranteed to always have at least one value by the clap configuration
-    let first_root = directories
-        .first()
-        .expect("directories is empty; this shouldn't be possible");
-
-    let overrides = OverrideBuilder::new(first_root)
-        // Don't process files inside of tools/typeshare/
-        .add("!**/tools/typeshare/**")
-        .context("Failed to parse override")?
-        .build()
-        .context("Failed to build override")?;
-
-    let mut walker_builder = WalkBuilder::new(first_root);
-    // Sort walker output for deterministic output across platforms
-    walker_builder
-        .sort_by_file_path(Path::cmp)
-        .types(types.build().context("Failed to build types")?)
-        .overrides(overrides)
-        .follow_links(options.follow_links);
-
-    for root in directories {
-        walker_builder.add(root);
-    }
-
     let destination = if let Some(ref file) = options.output.file {
         Output::File(file)
     } else if let Some(ref folder) = options.output.folder {
@@ -126,43 +111,70 @@ fn main() -> anyhow::Result<()> {
 
     let multi_file = matches!(destination, Output::Folder(_));
     let target_os = config.target_os.clone();
-
     let mut lang = language(language_type, config, multi_file);
-    let ignored_types = lang.ignored_reference_types();
 
-    // The walker ignores directories that are git-ignored. If you need
-    // a git-ignored directory to be processed, add the specific directory to
-    // the list of directories given to typeshare when it's invoked in the
-    // makefiles
-    // TODO: The `ignore` walker supports parallel walking. We should use this
-    // and implement a `ParallelVisitor` that builds up the mapping of parsed
-    // data. That way both walking and parsing are in parallel.
-    // https://docs.rs/ignore/latest/ignore/struct.WalkParallel.html
-    let crate_parsed_data = parse_input(
-        parser_inputs(walker_builder, language_type, multi_file).par_bridge(),
-        &ignored_types,
+    let parse_context = ParseContext {
+        ignored_types: lang.ignored_reference_types(),
         multi_file,
-        &target_os,
+        target_os,
+    };
+
+    let mut parsed_data = parallel_parse(
+        &parse_context,
+        walker_builder(directories, options)?,
+        language_type,
     )?;
+
+    reconcile_aliases(&mut parsed_data);
 
     // Collect all the types into a map of the file name they
     // belong too and the list of type names. Used for generating
     // imports in generated files.
     let import_candidates = if multi_file {
-        all_types(&crate_parsed_data)
+        all_types(&mut parsed_data)
     } else {
         HashMap::new()
     };
 
-    check_parse_errors(&crate_parsed_data)?;
-    write_generated(
-        destination,
-        lang.as_mut(),
-        crate_parsed_data,
-        import_candidates,
-    )?;
+    check_parse_errors(&parsed_data)?;
 
+    info!("typeshare started writing generated types");
+
+    write_generated(destination, lang.as_mut(), parsed_data, import_candidates)?;
+
+    info!("typeshare finished generating types");
     Ok(())
+}
+
+fn walker_builder(
+    directories: &[std::path::PathBuf],
+    options: &Args,
+) -> anyhow::Result<WalkBuilder> {
+    let mut types = TypesBuilder::new();
+    types
+        .add("rust", "*.rs")
+        .context("Failed to add rust type extensions")?;
+    types.select("rust");
+
+    let first_root = directories
+        .first()
+        .expect("directories is empty; this shouldn't be possible");
+    let overrides = OverrideBuilder::new(first_root)
+        // Don't process files inside of tools/typeshare/
+        .add("!**/tools/typeshare/**")
+        .context("Failed to parse override")?
+        .build()
+        .context("Failed to build override")?;
+    let mut walker_builder = WalkBuilder::new(first_root);
+    walker_builder
+        .sort_by_file_path(Path::cmp)
+        .types(types.build().context("Failed to build types")?)
+        .overrides(overrides)
+        .follow_links(options.follow_links);
+    for root in directories.iter().skip(1) {
+        walker_builder.add(root);
+    }
+    Ok(walker_builder)
 }
 
 /// Get the language trait impl for the given supported language and configuration.
@@ -255,7 +267,13 @@ fn override_configuration(mut config: Config, options: &Args) -> anyhow::Result<
         if let Some(go_package) = options.go_package.as_ref() {
             config.go.package = go_package.to_string();
         }
-        assert_go_package_present(&config)?;
+
+        if matches!(options.language, Some(args::AvailableLanguage::Go)) {
+            anyhow::ensure!(
+                    !config.go.package.is_empty(),
+                   "Please provide a package name in the typeshare.toml or using --go-package <package name>"
+                );
+        }
     }
 
     config.target_os = options.target_os.as_deref().unwrap_or_default().to_vec();
@@ -273,25 +291,16 @@ fn check_parse_errors(parsed_crates: &BTreeMap<CrateName, ParsedData>) -> anyhow
         errors_encountered = true;
         for error in &data.errors {
             error!(
-                "Parsing error: \"{}\" in crate \"{}\" for file \"{}\"",
-                error.error, error.crate_name, error.file_name
+                "Parsing error: \"{}\" in file \"{}\"",
+                error.error, error.file_name
             );
         }
     }
 
     if errors_encountered {
+        error!("Errors encountered during parsing.");
         Err(anyhow!("Errors encountered during parsing."))
     } else {
         Ok(())
     }
-}
-
-#[cfg(feature = "go")]
-fn assert_go_package_present(config: &Config) -> anyhow::Result<()> {
-    if config.go.package.is_empty() {
-        return Err(anyhow!(
-            "Please provide a package name in the typeshare.toml or using --go-package <package name>"
-        ));
-    }
-    Ok(())
 }