+++ cli/src/parse.rs
@@ -1,51 +1,48 @@
 //! Source file parsing.
 use anyhow::Context;
-use ignore::WalkBuilder;
-use rayon::iter::{IntoParallelIterator, ParallelIterator};
+use crossbeam::channel::bounded;
+use ignore::{DirEntry, WalkBuilder, WalkState};
 use std::{
-    collections::{hash_map::Entry, BTreeMap, HashMap},
-    path::PathBuf,
+    collections::{BTreeMap, HashMap},
+    mem, thread,
 };
 use typeshare_core::{
+    context::{ParseContext, ParseFileContext},
     language::{CrateName, CrateTypes, SupportedLanguage, SINGLE_FILE_CRATE_NAME},
-    parser::ParsedData,
+    parser::{ParseError, ParsedData},
     RenameExt,
 };
 
-/// Input data for parsing each source file.
-pub struct ParserInput {
-    /// Rust source file path.
-    file_path: PathBuf,
-    /// File name source from crate for output.
-    file_name: String,
-    /// The crate name the source file belongs to.
-    crate_name: CrateName,
-}
-
-/// Walk the source folder and collect all parser inputs.
-pub fn parser_inputs(
-    walker_builder: WalkBuilder,
-    language_type: SupportedLanguage,
+fn parse_file_context(
     multi_file: bool,
-) -> impl Iterator<Item = ParserInput> {
-    walker_builder
-        .build()
-        .filter_map(Result::ok)
-        .filter(|dir_entry| !dir_entry.path().is_dir())
-        .filter_map(move |dir_entry| {
-            let crate_name = if multi_file {
-                CrateName::find_crate_name(dir_entry.path())?
-            } else {
-                SINGLE_FILE_CRATE_NAME
-            };
-            let file_path = dir_entry.path().to_path_buf();
-            let file_name = output_file_name(language_type, &crate_name);
-            Some(ParserInput {
-                file_path,
-                file_name,
-                crate_name,
-            })
-        })
+    language_type: SupportedLanguage,
+    dir_entry: &DirEntry,
+) -> anyhow::Result<Option<ParseFileContext>> {
+    let crate_name = if multi_file {
+        let Some(crate_name) = CrateName::find_crate_name(dir_entry.path()) else {
+            return Ok(None);
+        };
+        crate_name
+    } else {
+        SINGLE_FILE_CRATE_NAME
+    };
+    let file_path = dir_entry.path().to_path_buf();
+    let out_file_name = output_file_name(language_type, &crate_name);
+
+    let input_file = file_path
+        .to_str()
+        .map(ToOwned::to_owned)
+        .unwrap_or_default();
+
+    let parse_file_context = ParseFileContext {
+        source_code: std::fs::read_to_string(&file_path)
+            .with_context(|| format!("Failed to read input: {input_file}"))?,
+        crate_name,
+        file_name: out_file_name,
+        file_path,
+    };
+
+    Ok(Some(parse_file_context))
 }
 
 /// The output file name to write to.
@@ -67,69 +64,84 @@ fn output_file_name(language_type: SupportedLanguage, crate_name: &CrateName) ->
 
 /// Collect all the typeshared types into a mapping of crate names to typeshared types. This
 /// mapping is used to lookup and generated import statements for generated files.
-pub fn all_types(file_mappings: &BTreeMap<CrateName, ParsedData>) -> CrateTypes {
+pub fn all_types(file_mappings: &mut BTreeMap<CrateName, ParsedData>) -> CrateTypes {
     file_mappings
-        .iter()
-        .map(|(crate_name, parsed_data)| (crate_name, parsed_data.type_names.clone()))
+        .iter_mut()
+        .map(|(crate_name, parsed_data)| (crate_name, mem::take(&mut parsed_data.type_names)))
         .fold(
             HashMap::new(),
             |mut import_map: CrateTypes, (crate_name, type_names)| {
-                match import_map.entry(crate_name.clone()) {
-                    Entry::Occupied(mut e) => {
-                        e.get_mut().extend(type_names);
-                    }
-                    Entry::Vacant(e) => {
-                        e.insert(type_names);
-                    }
-                }
+                import_map
+                    .entry(crate_name.clone())
+                    .or_default()
+                    .extend(type_names);
                 import_map
             },
         )
 }
 
-/// Collect all the parsed sources into a mapping of crate name to parsed data.
-pub fn parse_input(
-    inputs: impl ParallelIterator<Item = ParserInput>,
-    ignored_types: &[&str],
-    multi_file: bool,
-    target_os: &[String],
+fn parse_dir_entry(
+    parse_context: &ParseContext,
+    language_type: SupportedLanguage,
+    dir_entry: &DirEntry,
+) -> Result<Option<ParsedData>, ParseError> {
+    if dir_entry.path().is_dir() {
+        return Ok(None);
+    }
+
+    let Some(parse_file_context) =
+        parse_file_context(parse_context.multi_file, language_type, dir_entry)
+            .map_err(|err| ParseError::IOError(err.to_string()))?
+    else {
+        return Ok(None);
+    };
+
+    typeshare_core::parser::parse(parse_context, parse_file_context)
+}
+
+/// Use parallel builder to walk all source directories concurrently.
+pub fn parallel_parse(
+    parse_context: &ParseContext,
+    walker_builder: WalkBuilder,
+    language_type: SupportedLanguage,
 ) -> anyhow::Result<BTreeMap<CrateName, ParsedData>> {
-    inputs
-        .into_par_iter()
-        .try_fold(
-            BTreeMap::new,
-            |mut parsed_crates: BTreeMap<CrateName, ParsedData>,
-             ParserInput {
-                 file_path,
-                 file_name,
-                 crate_name,
-             }| {
-                let parsed_result = typeshare_core::parser::parse(
-                    &std::fs::read_to_string(&file_path)
-                        .with_context(|| format!("Failed to read input: {file_name}"))?,
-                    crate_name.clone(),
-                    file_name.clone(),
-                    file_path,
-                    ignored_types,
-                    multi_file,
-                    target_os,
-                )
-                .with_context(|| format!("Failed to parse: {file_name}"))?;
+    let (tx, rx) = bounded::<anyhow::Result<ParsedData>>(100);
 
-                if let Some(parsed_data) = parsed_result {
-                    parsed_crates
-                        .entry(crate_name)
-                        .or_default()
-                        .add(parsed_data);
+    let collector_thread = thread::spawn(move || {
+        let mut crate_parsed_data: BTreeMap<CrateName, ParsedData> = BTreeMap::new();
+
+        for result in rx {
+            let parsed_data = result?;
+            let crate_name = parsed_data.crate_name.clone();
+            // Append each yielded parsed data by its respective crate.
+            *crate_parsed_data.entry(crate_name).or_default() += parsed_data;
+        }
+
+        Ok(crate_parsed_data)
+    });
+
+    walker_builder.build_parallel().run(|| {
+        let tx = tx.clone();
+
+        Box::new(move |result| {
+            let result = result.context("Failed traversing").and_then(|dir_entry| {
+                parse_dir_entry(parse_context, language_type, &dir_entry)
+                    .with_context(|| format!("Parsing failed: {:?}", dir_entry.path()))
+            });
+            match result {
+                Ok(Some(parsed_data)) => {
+                    tx.send(Ok(parsed_data)).unwrap();
+                    WalkState::Continue
                 }
-
-                Ok(parsed_crates)
-            },
-        )
-        .try_reduce(BTreeMap::new, |mut file_maps, parsed_crates| {
-            for (crate_name, parsed_data) in parsed_crates {
-                file_maps.entry(crate_name).or_default().add(parsed_data);
+                Ok(None) => WalkState::Continue,
+                Err(err) => {
+                    tx.send(Err(err)).unwrap();
+                    WalkState::Quit
+                }
             }
-            Ok(file_maps)
         })
+    });
+
+    drop(tx);
+    collector_thread.join().unwrap()
 }