+++ core/src/language/mod.rs
@@ -90,7 +90,7 @@ pub type SortedCrateNames<'a> = &'a CrateName;
 /// A sorted type name ref.
 pub type SortedTypeNames<'a> = BTreeSet<&'a str>;
 
-/// Refence types by crate that are scoped for a given output module.
+/// Reference types by crate that are scoped for a given output module.
 pub type ScopedCrateTypes<'a> = BTreeMap<SortedCrateNames<'a>, SortedTypeNames<'a>>;
 
 /// All supported programming languages.
@@ -384,6 +384,7 @@ pub trait Language {
                     id: Id {
                         original: struct_name.clone(),
                         renamed: struct_name.clone(),
+                        serde_rename: false
                     },
                     fields: fields.clone(),
                     generic_types,